+++ src/main/java/wooteco/subway/line/repository/LineRepository.java
@@ -1,10 +1,10 @@
-package wooteco.subway.line.dao;
+package wooteco.subway.line.repository;
 
 import wooteco.subway.line.domain.Line;
 
 import java.util.List;
 
-public interface LineDao {
+public interface LineRepository {
 
     Line save(Line line);
