+++ src/main/java/wooteco/subway/line/service/LineService.java
@@ -1,8 +1,8 @@
 package wooteco.subway.line.service;
 
 import org.springframework.stereotype.Service;
-import wooteco.subway.line.dao.LineDao;
-import wooteco.subway.line.dao.dto.LineDto;
+import wooteco.subway.line.repository.LineRepository;
+import wooteco.subway.line.repository.dto.LineDto;
 import wooteco.subway.line.domain.Line;
 
 import java.util.ArrayList;
@@ -12,33 +12,33 @@ import java.util.stream.Collectors;
 @Service
 public class LineService {
 
-    private LineDao lineDao;
+    private LineRepository lineRepository;
 
-    public LineService(LineDao lineDao) {
-        this.lineDao = lineDao;
+    public LineService(LineRepository lineRepository) {
+        this.lineRepository = lineRepository;
     }
 
     public LineDto save(LineDto linedto) {
         Line line = new Line(linedto.getName(), linedto.getColor());
-        return LineDto.from(lineDao.save(line));
+        return LineDto.from(lineRepository.save(line));
     }
 
     public List<LineDto> findAll() {
-        return lineDao.findAll().stream()
+        return lineRepository.findAll().stream()
                 .map(LineDto::from)
                 .collect(Collectors.toList());
     }
 
     public LineDto findById(Long id) {
-        return LineDto.from(lineDao.findById(id));
+        return LineDto.from(lineRepository.findById(id));
     }
 
     public void delete(Long id) {
-        lineDao.delete(id);
+        lineRepository.delete(id);
     }
 
     public void update(LineDto lineDto) {
         Line line = new Line(lineDto.getId(), lineDto.getName(), lineDto.getColor(), new ArrayList<>());
-        lineDao.update(line);
+        lineRepository.update(line);
     }
 }
